- io: stream buffers  - refactoring

+++ source/utility/io/implementation/platform/freebsd/listener.h
@@ -11,7 +11,6 @@
 #include "utility/io/listener.h"
 
 #include <sys/event.h>
-
 #include <unordered_map>
 
 
@@ -29,16 +28,16 @@ public:
     void initialize()   override;
     void finalize()     override;
 
-    void add(IStream::TSharedPtr const &stream) override;
-    void del(IStream::TSharedPtr const &stream) override;
-    TEvents wait()                              override;
+    void    add(IStream::TSharedPtr const &stream) override;
+    void    del(IStream::TSharedPtr const &stream) override;
+    TEvents wait() override;
 
 private:
     std::vector<struct kevent>  m_events;
     std::vector<struct kevent>  m_monitor_events;
     int                         m_monitor_events_used_count;
     int                         m_kqueue;
-    std::unordered_map<int, IStream::TSharedPtr> m_map_fd_stream;
+    std::unordered_map<uintptr_t, IStream::TSharedPtr> m_map_fd_stream;
 };
